is it because of different sizes or because of a call to truncate?
nukclear_test and update_test crash when double clicked but work from
the command line, is this because of permissions with the memory mapped
file? are the paths different? should the files be put in
/tmp/var/simdb/ ?

simdb_15_test added to ignore list - it is the memory mapped file of
the database